requirements.txt
livekit-agents
python-dotenv
aiohttp
certifi
llama-index
llama-index-embeddings-huggingface
sentence-transformers
torch
transformers 